Rename objectutil to netvariablesutil
